add Coveralls support with new /coveralls endpoint
Coveralls (http://coveralls.io/) can send a webhook when it has new coverage data for a change: https://coveralls.zendesk.com/hc/en-us/articles/201774895-Notifications

this adds support for that with a new /coveralls endpoint. it's mostly the same as the /hook/... endpoint, and reuses most of the code, but looks up the GitHub PR based on the commit hash.

it also mildly refactors GithubBot.process_hook a bit to allow unset build_id, since Coveralls doesn't send one in its webhook.