add new boolean GH_COMMENT env var for whether to post GitHub comments
this is useful because we just use Shamer as a leaderboard that aggregates coverage contributions by user. we use Coveralls to collect and track coverage over time, which is great, so we don't need Shamer for that. setting this new env var to false lets us skip the GitHub comments, since Coveralls is already a status check on our GitHub PRs, and also lets us avoid storing coverage data in AWS at all.